refactor: standardize file naming conventions to PascalCase
- Rename all React components (.tsx) to PascalCase
- Rename all Astro components (.astro) to PascalCase
- Update all import statements across codebase
- Maintain pages in camelCase as per requirements
- Ensure all quality checks pass (type-check, lint, tests)
