+++ src/components/features/legal/LegalContent.astro
@@ -1,7 +1,7 @@
 ---
 // Use new i18n system from middleware
 const { t } = Astro.locals
-import GradientCircle from '@/components/common/gradient-circle.astro'
+import GradientCircle from '@/components/common/GradientCircle.astro'
 
 export interface Props {
 	title: string
