+++ src/components/features/legal/Section.astro
