+++ src/components/features/marketing/hero/HeroContent.astro
@@ -1,5 +1,5 @@
 ---
-import { ButtonWithLink } from '@/components/ui/forms/button-with-link'
+import { ButtonWithLink } from '@/components/ui/forms/ButtonWithLink'
 import { cn } from '@/lib/core/utils'
 
 // Use new i18n system from middleware
