+++ src/components/features/marketing/hero/Welcome.astro
@@ -1,9 +1,9 @@
 ---
 import { cn } from '@/lib/core/utils'
-import HeroContent from './hero-content.astro'
-import ScrollIndicator from '../interaction/scroll-indicator.astro'
-import MobileSidePanel from '../panels/mobile-side-panel.astro'
-import SidePanel from '../panels/side-panel.astro'
+import HeroContent from './HeroContent.astro'
+import ScrollIndicator from '../interaction/ScrollIndicator.astro'
+import MobileSidePanel from '../panels/MobileSidePanel.astro'
+import SidePanel from '../panels/SidePanel.astro'
 ---
 
 <!-- Hero Section -->
