+++ src/components/features/marketing/panels/MobileSidePanel.astro
@@ -1,6 +1,6 @@
 ---
 import { cn } from '@/lib/core/utils'
-import SidePanel from './side-panel.astro'
+import SidePanel from './SidePanel.astro'
 ---
 
 <!-- Mobile Side Panel - Appears after scroll -->
