+++ src/components/features/marketing/panels/SidePanel.astro
