+++ src/components/features/pricing/PricingCard.astro
@@ -4,7 +4,7 @@ import { Check } from 'lucide-react'
 // Use new i18n system from middleware
 const { t } = Astro.locals
 import { translatePlan } from '@/lib/i18n/translate-utils'
-import { Button } from '@/components/ui/forms/button'
+import { Button } from '@/components/ui/forms/Button'
 import {
 	PLAN_THEMES,
 	PRICING_CARD_CLASSES,
