+++ src/components/features/pricing/PricingHero.astro
@@ -2,8 +2,8 @@
 import { cn } from '@/lib/core/utils'
 // Use new i18n system from middleware
 const { t } = Astro.locals
-import GradientCircle from '@/components/common/gradient-circle.astro'
-import { ButtonWithLink } from '@/components/ui/forms/button-with-link'
+import GradientCircle from '@/components/common/GradientCircle.astro'
+import { ButtonWithLink } from '@/components/ui/forms/ButtonWithLink'
 ---
 
 <section class="relative overflow-hidden">
