+++ src/components/features/workflow/cards/StepCard.astro
@@ -4,7 +4,7 @@ const { t } = Astro.locals
 import { translateSteps } from '@/lib/i18n/translate-utils'
 import { getStepDesign } from '../step-designs'
 import { VARIANT_FEATURES, getVariantClasses } from './step-card-variants'
-import PatternBackground from '@/components/ui/patterns/pattern-background.astro'
+import PatternBackground from '@/components/ui/patterns/PatternBackground.astro'
 import type { StepCardVariant } from './step-card-variants'
 import type { StepKey } from '@/types/i18n'
