+++ src/components/layout/header/Controls.astro
@@ -1,6 +1,6 @@
 ---
-import DesktopControls from './desktop/controls-desktop.astro'
-import MobileControls from './mobile/controls-mobile.astro'
+import DesktopControls from './desktop/ControlsDesktop.astro'
+import MobileControls from './mobile/ControlsMobile.astro'
 
 interface Props {
 	variant?: 'desktop' | 'mobile'
