+++ src/components/layout/header/Header.astro
@@ -1,9 +1,9 @@
 ---
-import Navigation from './navigation.astro'
-import Controls from './controls.astro'
-import Logo from './logo.astro'
-import MobileButton from './mobile/mobile-button.astro'
-import MobileMenu from './mobile/mobile-menu.astro'
+import Navigation from './Navigation.astro'
+import Controls from './Controls.astro'
+import Logo from './Logo.astro'
+import MobileButton from './mobile/MobileButton.astro'
+import MobileMenu from './mobile/MobileMenu.astro'
 import { navigationLinks } from '@/i18n/config'
 ---
