+++ src/components/layout/header/mobile/MobileMenu.astro
@@ -1,8 +1,8 @@
 ---
-import Navigation from '../navigation.astro'
-import Controls from '../controls.astro'
-import Logo from '../logo.astro'
-import MobileButton from './mobile-button.astro'
+import Navigation from '../Navigation.astro'
+import Controls from '../Controls.astro'
+import Logo from '../Logo.astro'
+import MobileButton from './MobileButton.astro'
 import type { NavigationLinks } from '@/types/i18n'
 
 interface Props {
