+++ src/components/ui/patterns/PatternBackground.astro
@@ -4,7 +4,7 @@
  * Provides a simple API for adding decorative SVG patterns to components
  */
 
-import SVGPatternGenerator from './svg-pattern-generator.astro'
+import SVGPatternGenerator from './SvgPatternGenerator.astro'
 import { cn } from '@/lib/core/utils'
 
 export interface Props {
