+++ src/components/ui/patterns/SvgPatternGenerator.astro
