+++ src/layouts/BaseLayout.astro
@@ -3,7 +3,7 @@ import '../styles/global.css'
 import '../styles/base.css'
 import '../styles/animations.css'
 import '../styles/liquid-effects.css'
-import { ContextProvider } from '../lib/i18n/context-provider'
+import { ContextProvider } from '../lib/i18n/ContextProvider'
 
 export interface Props {
 	title: string
