+++ src/lib/i18n/ContextProvider.tsx
@@ -6,7 +6,7 @@
 import type { PropsWithChildren, ReactNode } from 'react'
 
 import { DEFAULT_LOCALE, SUPPORTED_LOCALES } from '../../i18n/config'
-import { I18nProvider } from './react'
+import { I18nProvider } from './I18nReact'
 
 import type { Locale } from '@/types/i18n'
