+++ src/pages/500.astro
@@ -1,7 +1,7 @@
 ---
 import BaseLayout from '@/layouts/BaseLayout.astro'
-import Header from '@/components/layout/header/header.astro'
-import Footer from '@/components/layout/footer/footer.astro'
+import Header from '@/components/layout/header/Header.astro'
+import Footer from '@/components/layout/footer/Footer.astro'
 import { getRelativeLocaleUrl } from 'astro:i18n'
 // Use new i18n system from middleware
 const { locale: currentLocale, t } = Astro.locals
