+++ src/pages/[locale]/how-we-work.astro
@@ -1,9 +1,9 @@
 ---
 import BaseLayout from '@/layouts/BaseLayout.astro'
-import Header from '@/components/layout/header/header.astro'
-import Footer from '@/components/layout/footer/footer.astro'
-import GradientCircle from '@/components/common/gradient-circle.astro'
-import WorkflowJourney from '@/components/features/workflow/interactive/workflow-journey.astro'
+import Header from '@/components/layout/header/Header.astro'
+import Footer from '@/components/layout/footer/Footer.astro'
+import GradientCircle from '@/components/common/GradientCircle.astro'
+import WorkflowJourney from '@/components/features/workflow/interactive/WorkflowJourney.astro'
 import MobileTimeline from '@/components/features/workflow/timeline/MobileTimeline.astro'
 // Use new i18n system from middleware
 const { t } = Astro.locals
