+++ src/pages/[locale]/index.astro
@@ -2,11 +2,11 @@
 // Welcome to Astro! Everything between these triple-dash code fences
 // is your "component frontmatter". It never runs in the browser.
 
-import Welcome from '@/components/features/marketing/hero/welcome.astro'
+import Welcome from '@/components/features/marketing/hero/Welcome.astro'
 import BaseLayout from '@/layouts/BaseLayout.astro'
-import GradientCircle from '@/components/common/gradient-circle.astro'
-import Header from '@/components/layout/header/header.astro'
-import Footer from '@/components/layout/footer/footer.astro'
+import GradientCircle from '@/components/common/GradientCircle.astro'
+import Header from '@/components/layout/header/Header.astro'
+import Footer from '@/components/layout/footer/Footer.astro'
 import { cn } from '@/lib/core/utils'
 
 // i18n is now handled by middleware - no manual initialization needed
