+++ src/pages/[locale]/pricing.astro
@@ -1,8 +1,8 @@
 ---
 import BaseLayout from '@/layouts/BaseLayout.astro'
-import Header from '@/components/layout/header/header.astro'
-import Footer from '@/components/layout/footer/footer.astro'
-import GradientCircle from '@/components/common/gradient-circle.astro'
+import Header from '@/components/layout/header/Header.astro'
+import Footer from '@/components/layout/footer/Footer.astro'
+import GradientCircle from '@/components/common/GradientCircle.astro'
 import PricingHero from '@/components/features/pricing/PricingHero.astro'
 import PricingCard from '@/components/features/pricing/PricingCard.astro'
 import { cn } from '@/lib/core/utils'
