+++ src/pages/[locale]/terms.astro
@@ -1,9 +1,9 @@
 ---
 import BaseLayout from '@/layouts/BaseLayout.astro'
-import Header from '@/components/layout/header/header.astro'
-import Footer from '@/components/layout/footer/footer.astro'
-import LegalContent from '@/components/features/legal/legal-content.astro'
-import Section from '@/components/features/legal/section.astro'
+import Header from '@/components/layout/header/Header.astro'
+import Footer from '@/components/layout/footer/Footer.astro'
+import LegalContent from '@/components/features/legal/LegalContent.astro'
+import Section from '@/components/features/legal/Section.astro'
 // Use new i18n system from middleware
 const { t } = Astro.locals
